third_party/WebKit/LayoutTests/animations/svg-attribute-composition/svg-gradientTransform-composition-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
This is a testharness.js-based test.
PASS This test uses interpolation-test.js. 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (0) is [translate(10,10) translate(0,10)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (0.2) is [translate(10,10) translate(4,18)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (0.6) is [translate(10,10) translate(12,34)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (1) is [translate(10,10) translate(20,50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (-0.4) is [translate(10,10) translate(-8,-6)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , - 8 , - 6 " but got "2 , 0 , 1 , 0 , 0 , 1 , - 8 , - 6 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (0) is [translate(10,10) translate(0,10)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 0 , 10 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 10 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (0.2) is [translate(10,10) translate(4,18)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 4 , 18 " but got "2 , 0 , 1 , 0 , 0 , 1 , 4 , 18 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (0.6) is [translate(10,10) translate(12,34)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 12 , 34 " but got "2 , 0 , 1 , 0 , 0 , 1 , 12 , 34 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (1) is [translate(10,10) translate(20,50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 20 , 50 " but got "2 , 0 , 1 , 0 , 0 , 1 , 20 , 50 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from add [translate(0 10)] to add [translate(20 50)] at (1.4) is [translate(10,10) translate(28,66)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 28 , 66 " but got "2 , 0 , 1 , 0 , 0 , 1 , 28 , 66 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from replace [translate(10 20)] to add [translate(20 50)] at (-0.4) is [translate(10 20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 " but got "2 , 0 , 1 , 0 , 0 , 1 , 6 , 8 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from replace [translate(10 20)] to add [translate(20 50)] at (0) is [translate(10 20)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from replace [translate(10 20)] to add [translate(20 50)] at (0.2) is [translate(10 20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 " but got "2 , 0 , 1 , 0 , 0 , 1 , 12 , 26 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from replace [translate(10 20)] to add [translate(20 50)] at (0.6) is [translate(10 10) translate(20 50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 20 , 50 " but got "2 , 0 , 1 , 0 , 0 , 1 , 16 , 38 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from replace [translate(10 20)] to add [translate(20 50)] at (1) is [translate(10 10) translate(20 50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 20 , 50 " but got "2 , 0 , 1 , 0 , 0 , 1 , 20 , 50 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 10)] from replace [translate(10 20)] to add [translate(20 50)] at (1.4) is [translate(10 10) translate(20 50)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 10 , 2 , 0 , 1 , 0 , 0 , 1 , 20 , 50 " but got "2 , 0 , 1 , 0 , 0 , 1 , 24 , 62 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 20)] from neutral to replace [translate(30 60)] at (-0.4) is [translate(2,4)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 2 , 4 " but got "2 , 0 , 1 , 0 , 0 , 1 , - 12 , - 24 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 20)] from neutral to replace [translate(30 60)] at (0) is [translate(10,20)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 10 , 20 " but got "2 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 20)] from neutral to replace [translate(30 60)] at (0.2) is [translate(14,28)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 14 , 28 " but got "2 , 0 , 1 , 0 , 0 , 1 , 6 , 12 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 20)] from neutral to replace [translate(30 60)] at (0.6) is [translate(22,44)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 22 , 44 " but got "2 , 0 , 1 , 0 , 0 , 1 , 18 , 36 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 20)] from neutral to replace [translate(30 60)] at (1) is [translate(30,60)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [translate(10 20)] from neutral to replace [translate(30 60)] at (1.4) is [translate(38,76)] assert_equals: expected "2 , 0 , 1 , 0 , 0 , 1 , 38 , 76 " but got "2 , 0 , 1 , 0 , 0 , 1 , 42 , 84 "
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (0) is [scale(10 5) scale(10 10)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (0.2) is [scale(10 5) scale(12 18)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (0.6) is [scale(10 5) scale(16 34)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (1) is [scale(10 5) scale(20 50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (-0.4) is [scale(10 5) scale(6 -6)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 6 , 0 , 0 , - 6 , 0 , 0 " but got "3 , 0 , 6 , 0 , 0 , - 6 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (0) is [scale(10 5) scale(10 10)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 " but got "3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (0.2) is [scale(10 5) scale(12 18)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 12 , 0 , 0 , 18 , 0 , 0 " but got "3 , 0 , 12 , 0 , 0 , 18 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (0.6) is [scale(10 5) scale(16 34)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 16 , 0 , 0 , 34 , 0 , 0 " but got "3 , 0 , 16 , 0 , 0 , 34 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (1) is [scale(10 5) scale(20 50)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 " but got "3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from add [scale(10 10)] to add [scale(20 50)] at (1.4) is [scale(10 5) scale(24 66)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 24 , 0 , 0 , 66 , 0 , 0 " but got "3 , 0 , 24 , 0 , 0 , 66 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from replace [scale(10 10)] to add [scale(20 50)] at (-0.4) is [scale(10 10)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 " but got "3 , 0 , 6 , 0 , 0 , - 6 , 0 , 0 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from replace [scale(10 10)] to add [scale(20 50)] at (0) is [scale(10 10)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from replace [scale(10 10)] to add [scale(20 50)] at (0.2) is [scale(10 10)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 " but got "3 , 0 , 12 , 0 , 0 , 18 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from replace [scale(10 10)] to add [scale(20 50)] at (0.6) is [scale(10 5) scale(20 50)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 " but got "3 , 0 , 16 , 0 , 0 , 34 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from replace [scale(10 10)] to add [scale(20 50)] at (1) is [scale(10 5) scale(20 50)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 " but got "3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 5)] from replace [scale(10 10)] to add [scale(20 50)] at (1.4) is [scale(10 5) scale(20 50)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 5 , 0 , 0 , 3 , 0 , 20 , 0 , 0 , 50 , 0 , 0 " but got "3 , 0 , 24 , 0 , 0 , 66 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 10)] from neutral to replace [scale(20 50)] at (-0.4) is [scale(6 -6)] assert_equals: expected "3 , 0 , 6 , 0 , 0 , - 6 , 0 , 0 " but got "3 , 0 , - 8 , 0 , 0 , - 20 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 10)] from neutral to replace [scale(20 50)] at (0) is [scale(10 10)] assert_equals: expected "3 , 0 , 10 , 0 , 0 , 10 , 0 , 0 " but got "3 , 0 , 0 , 0 , 0 , 0 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 10)] from neutral to replace [scale(20 50)] at (0.2) is [scale(12 18)] assert_equals: expected "3 , 0 , 12 , 0 , 0 , 18 , 0 , 0 " but got "3 , 0 , 4 , 0 , 0 , 10 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 10)] from neutral to replace [scale(20 50)] at (0.6) is [scale(16 34)] assert_equals: expected "3 , 0 , 16 , 0 , 0 , 34 , 0 , 0 " but got "3 , 0 , 12 , 0 , 0 , 30 , 0 , 0 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 10)] from neutral to replace [scale(20 50)] at (1) is [scale(20 50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [scale(10 10)] from neutral to replace [scale(20 50)] at (1.4) is [scale(24 66)] assert_equals: expected "3 , 0 , 24 , 0 , 0 , 66 , 0 , 0 " but got "3 , 0 , 28 , 0 , 0 , 70 , 0 , 0 "
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (0) is [rotate(10 10 10) rotate(10 20 30)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (0.2) is [rotate(10 10 10) rotate(16 32 56)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (0.6) is [rotate(10 10 10) rotate(28 56 108)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (1) is [rotate(10 10 10) rotate(40 80 160)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (-0.4) is [rotate(10 10 10) rotate(-2 -4 -22)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , - 2 , 1 , - 0.03 , 0.03 , 1 , 0.77 , - 0.15 " but got "4 , - 2 , 1 , - 0.03 , 0.03 , 1 , 0.77 , - 0.15 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (0) is [rotate(10 10 10) rotate(10 20 30)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (0.2) is [rotate(10 10 10) rotate(16 32 56)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 16 , 0.96 , 0.28 , - 0.28 , 0.96 , 16.68 , - 6.65 " but got "4 , 16 , 0.96 , 0.28 , - 0.28 , 0.96 , 16.68 , - 6.65 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (0.6) is [rotate(10 10 10) rotate(28 56 108)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 28 , 0.88 , 0.47 , - 0.47 , 0.88 , 57.26 , - 13.65 " but got "4 , 28 , 0.88 , 0.47 , - 0.47 , 0.88 , 57.26 , - 13.65 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (1) is [rotate(10 10 10) rotate(40 80 160)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from add [rotate(10 20 30)] to add [rotate(40 80 160)] at (1.4) is [rotate(10 10 10) rotate(52 104 212)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 52 , 0.62 , 0.79 , - 0.79 , 0.62 , 207.03 , - 0.47 " but got "4 , 52 , 0.62 , 0.79 , - 0.79 , 0.62 , 207.03 , - 0.47 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from replace [rotate(10 20 30)] to add [rotate(40 80 160)] at (-0.4) is [rotate(10 20 30)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "4 , - 2 , 1 , - 0.03 , 0.03 , 1 , 0.77 , - 0.15 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from replace [rotate(10 20 30)] to add [rotate(40 80 160)] at (0) is [rotate(10 20 30)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from replace [rotate(10 20 30)] to add [rotate(40 80 160)] at (0.2) is [rotate(10 20 30)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "4 , 16 , 0.96 , 0.28 , - 0.28 , 0.96 , 16.68 , - 6.65 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from replace [rotate(10 20 30)] to add [rotate(40 80 160)] at (0.6) is [rotate(10 10 10) rotate(40 80 160)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "4 , 28 , 0.88 , 0.47 , - 0.47 , 0.88 , 57.26 , - 13.65 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from replace [rotate(10 20 30)] to add [rotate(40 80 160)] at (1) is [rotate(10 10 10) rotate(40 80 160)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 10 10)] from replace [rotate(10 20 30)] to add [rotate(40 80 160)] at (1.4) is [rotate(10 10 10) rotate(40 80 160)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 1.89 , - 1.58 , 4 , 40 , 0.77 , 0.64 , - 0.64 , 0.77 , 121.56 , - 13.99 " but got "4 , 52 , 0.62 , 0.79 , - 0.79 , 0.62 , 207.03 , - 0.47 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 20 30)] from neutral to replace [rotate(40 80 160)] at (-0.4) is [rotate(-2 -4 -22)] assert_equals: expected "4 , - 2 , 1 , - 0.03 , 0.03 , 1 , 0.77 , - 0.15 " but got "4 , - 16 , 0.96 , - 0.28 , 0.28 , 0.96 , 16.4 , - 11.3 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 20 30)] from neutral to replace [rotate(40 80 160)] at (0) is [rotate(10 20 30)] assert_equals: expected "4 , 10 , 0.98 , 0.17 , - 0.17 , 0.98 , 5.51 , - 3.02 " but got "4 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 20 30)] from neutral to replace [rotate(40 80 160)] at (0.2) is [rotate(16 32 56)] assert_equals: expected "4 , 16 , 0.96 , 0.28 , - 0.28 , 0.96 , 16.68 , - 6.65 " but got "4 , 8 , 0.99 , 0.14 , - 0.14 , 0.99 , 4.61 , - 1.92 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 20 30)] from neutral to replace [rotate(40 80 160)] at (0.6) is [rotate(28 56 108)] assert_equals: expected "4 , 28 , 0.88 , 0.47 , - 0.47 , 0.88 , 57.26 , - 13.65 " but got "4 , 24 , 0.91 , 0.41 , - 0.41 , 0.91 , 43.2 , - 11.22 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 20 30)] from neutral to replace [rotate(40 80 160)] at (1) is [rotate(40 80 160)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [rotate(10 20 30)] from neutral to replace [rotate(40 80 160)] at (1.4) is [rotate(52 104 212)] assert_equals: expected "4 , 52 , 0.62 , 0.79 , - 0.79 , 0.62 , 207.03 , - 0.47 " but got "4 , 56 , 0.56 , 0.83 , - 0.83 , 0.56 , 235.07 , 5.89 "
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (0) is [skewX(10) skewX(30)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (0.2) is [skewX(10) skewX(34)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (0.6) is [skewX(10) skewX(42)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (1) is [skewX(10) skewX(50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (-0.4) is [skewX(10) skewX(22)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 22 , 1 , 0 , 0.4 , 1 , 0 , 0 " but got "5 , 22 , 1 , 0 , 0.4 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (0) is [skewX(10) skewX(30)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 30 , 1 , 0 , 0.58 , 1 , 0 , 0 " but got "5 , 30 , 1 , 0 , 0.58 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (0.2) is [skewX(10) skewX(34)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 34 , 1 , 0 , 0.67 , 1 , 0 , 0 " but got "5 , 34 , 1 , 0 , 0.67 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (0.6) is [skewX(10) skewX(42)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 42 , 1 , 0 , 0.9 , 1 , 0 , 0 " but got "5 , 42 , 1 , 0 , 0.9 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (1) is [skewX(10) skewX(50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 50 , 1 , 0 , 1.19 , 1 , 0 , 0 " but got "5 , 50 , 1 , 0 , 1.19 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from add [skewX(30)] to add [skewX(50)] at (1.4) is [skewX(10) skewX(58)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 58 , 1 , 0 , 1.6 , 1 , 0 , 0 " but got "5 , 58 , 1 , 0 , 1.6 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from replace [skewX(30)] to add [skewX(50)] at (-0.4) is [skewX(30)] assert_equals: expected "5 , 30 , 1 , 0 , 0.58 , 1 , 0 , 0 " but got "5 , 22 , 1 , 0 , 0.4 , 1 , 0 , 0 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from replace [skewX(30)] to add [skewX(50)] at (0) is [skewX(30)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from replace [skewX(30)] to add [skewX(50)] at (0.2) is [skewX(30)] assert_equals: expected "5 , 30 , 1 , 0 , 0.58 , 1 , 0 , 0 " but got "5 , 34 , 1 , 0 , 0.67 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from replace [skewX(30)] to add [skewX(50)] at (0.6) is [skewX(10) skewX(50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 50 , 1 , 0 , 1.19 , 1 , 0 , 0 " but got "5 , 42 , 1 , 0 , 0.9 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from replace [skewX(30)] to add [skewX(50)] at (1) is [skewX(10) skewX(50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 50 , 1 , 0 , 1.19 , 1 , 0 , 0 " but got "5 , 50 , 1 , 0 , 1.19 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(10)] from replace [skewX(30)] to add [skewX(50)] at (1.4) is [skewX(10) skewX(50)] assert_equals: expected "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 , 5 , 50 , 1 , 0 , 1.19 , 1 , 0 , 0 " but got "5 , 58 , 1 , 0 , 1.6 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(30)] from neutral to replace [skewX(50)] at (-0.4) is [skewX(22)] assert_equals: expected "5 , 22 , 1 , 0 , 0.4 , 1 , 0 , 0 " but got "5 , - 20 , 1 , 0 , - 0.36 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(30)] from neutral to replace [skewX(50)] at (0) is [skewX(30)] assert_equals: expected "5 , 30 , 1 , 0 , 0.58 , 1 , 0 , 0 " but got "5 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(30)] from neutral to replace [skewX(50)] at (0.2) is [skewX(34)] assert_equals: expected "5 , 34 , 1 , 0 , 0.67 , 1 , 0 , 0 " but got "5 , 10 , 1 , 0 , 0.18 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(30)] from neutral to replace [skewX(50)] at (0.6) is [skewX(42)] assert_equals: expected "5 , 42 , 1 , 0 , 0.9 , 1 , 0 , 0 " but got "5 , 30 , 1 , 0 , 0.58 , 1 , 0 , 0 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(30)] from neutral to replace [skewX(50)] at (1) is [skewX(50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewX(30)] from neutral to replace [skewX(50)] at (1.4) is [skewX(58)] assert_equals: expected "5 , 58 , 1 , 0 , 1.6 , 1 , 0 , 0 " but got "5 , 70 , 1 , 0 , 2.75 , 1 , 0 , 0 "
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (0) is [skewY(10) skewY(30)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (0.2) is [skewY(10) skewY(34)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (0.6) is [skewY(10) skewY(42)] 
PASS SMIL: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (1) is [skewY(10) skewY(50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (-0.4) is [skewY(10) skewY(22)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 22 , 1 , 0.4 , 0 , 1 , 0 , 0 " but got "6 , 22 , 1 , 0.4 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (0) is [skewY(10) skewY(30)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 30 , 1 , 0.58 , 0 , 1 , 0 , 0 " but got "6 , 30 , 1 , 0.58 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (0.2) is [skewY(10) skewY(34)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 34 , 1 , 0.67 , 0 , 1 , 0 , 0 " but got "6 , 34 , 1 , 0.67 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (0.6) is [skewY(10) skewY(42)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 42 , 1 , 0.9 , 0 , 1 , 0 , 0 " but got "6 , 42 , 1 , 0.9 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (1) is [skewY(10) skewY(50)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 50 , 1 , 1.19 , 0 , 1 , 0 , 0 " but got "6 , 50 , 1 , 1.19 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from add [skewY(30)] to add [skewY(50)] at (1.4) is [skewY(10) skewY(58)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 58 , 1 , 1.6 , 0 , 1 , 0 , 0 " but got "6 , 58 , 1 , 1.6 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from replace [skewY(30)] to add [skewY(50)] at (-0.4) is [skewY(30)] assert_equals: expected "6 , 30 , 1 , 0.58 , 0 , 1 , 0 , 0 " but got "6 , 22 , 1 , 0.4 , 0 , 1 , 0 , 0 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from replace [skewY(30)] to add [skewY(50)] at (0) is [skewY(30)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from replace [skewY(30)] to add [skewY(50)] at (0.2) is [skewY(30)] assert_equals: expected "6 , 30 , 1 , 0.58 , 0 , 1 , 0 , 0 " but got "6 , 34 , 1 , 0.67 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from replace [skewY(30)] to add [skewY(50)] at (0.6) is [skewY(10) skewY(50)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 50 , 1 , 1.19 , 0 , 1 , 0 , 0 " but got "6 , 42 , 1 , 0.9 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from replace [skewY(30)] to add [skewY(50)] at (1) is [skewY(10) skewY(50)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 50 , 1 , 1.19 , 0 , 1 , 0 , 0 " but got "6 , 50 , 1 , 1.19 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(10)] from replace [skewY(30)] to add [skewY(50)] at (1.4) is [skewY(10) skewY(50)] assert_equals: expected "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 , 6 , 50 , 1 , 1.19 , 0 , 1 , 0 , 0 " but got "6 , 58 , 1 , 1.6 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(30)] from neutral to replace [skewY(50)] at (-0.4) is [skewY(22)] assert_equals: expected "6 , 22 , 1 , 0.4 , 0 , 1 , 0 , 0 " but got "6 , - 20 , 1 , - 0.36 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(30)] from neutral to replace [skewY(50)] at (0) is [skewY(30)] assert_equals: expected "6 , 30 , 1 , 0.58 , 0 , 1 , 0 , 0 " but got "6 , 0 , 1 , 0 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(30)] from neutral to replace [skewY(50)] at (0.2) is [skewY(34)] assert_equals: expected "6 , 34 , 1 , 0.67 , 0 , 1 , 0 , 0 " but got "6 , 10 , 1 , 0.18 , 0 , 1 , 0 , 0 "
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(30)] from neutral to replace [skewY(50)] at (0.6) is [skewY(42)] assert_equals: expected "6 , 42 , 1 , 0.9 , 0 , 1 , 0 , 0 " but got "6 , 30 , 1 , 0.58 , 0 , 1 , 0 , 0 "
PASS Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(30)] from neutral to replace [skewY(50)] at (1) is [skewY(50)] 
FAIL Web Animations: Interpolate attribute <gradientTransform> with underlying [skewY(30)] from neutral to replace [skewY(50)] at (1.4) is [skewY(58)] assert_equals: expected "6 , 58 , 1 , 1.6 , 0 , 1 , 0 , 0 " but got "6 , 70 , 1 , 2.75 , 0 , 1 , 0 , 0 "
Harness: the test ran to completion.
